Added ShipmentsController and shipment_list template
The controller has a URL mapping to show a list of the
shipments. Mapping is on URL '/', since the list is the
entry point of the web app.

Upon GET requests on /, the controller recovers the list
of all shipments on the ShipmentRepository and returns
the template name for the shipment list.

Template iterates shipments and shows each shipment's name
in a link, or shows a message if no shipments are stored