Add 'new shipment' form
Created a static HTML with a form to add a new shipment.
So far the only input is the shipment name.

Added a link to the new form HTML file from the shipment
list page.

Controller processes POST requests to /shipment URL.
Shipment instance is captured thanks to the spring
framework magic, with the @ModelAttribute annotation.
Controller saves the shipment instance in the repository,
and then the application shows a page with the detail
of the new shipment.