Support shipments with one item in new_shipment form
Added one field in the form, name has to be "items[0]" since
items is an array inside shipment. The shipment instance passed
to the controller method already contains the item.

Before saving the shipment instance, we have to traverse its items
to save each one in the items repository, otherwise we get an invalid
transient state exception.