Javascript to add item field to new shipment form
Added a button to add new field in the form.

Added the item fields to an identified div that can be referred
from the javascript. Addded classes to the span and input elements.

"Add item" button triggers a javascript that clones the last field
div container, modifying the text and the "name" attribute of the
input element (it must contain an index that needs to be incremented)
and resetting the input text. The cloned field is added to the
id'ied div.