Add remove buttons to item fields
Added "Remove" button next to each field.

Added function to define the behavior. It receives the index
of the item to remove. Since the array of items has to be consistent,
the function loops through the next items in the array, decreasing
their index. The index shows up in the field text, in the name attribute
of the input element, and in the parameter of the onclick function
of the remove element.

Renamed the javascript file to something more accurate.