Disable 'Submit' button when any of the fields is empty
Use a jquery statement to "loop" the input fields that are
required and check wether they are empty or not. If any of them
is empty the submit button is disabled. When all of them are not
empty, button is enabled.

This logic is binded to "oninput" events in the text input fields,
and also to the 'Add item' button, as it add a new empty field and
therefore the 'Submit' has to be disabled.