Delete shipments
Added handler in ShipmentController, listening to DELETE requests
in URL "/shipment/{id}". The handler removes a shipment with a
given id from the shipment repository, and all its items from
the items repository.

Added Delete button in shipment_detail template. Since HTML forms
are limited to PUT and GET requests, the functionality to the button
is added in a javascript function that performs an asynchronous
request with jquery ajax. Upon successful delete request, the user
is redirected to the shipment list.

We do not want the server to render any HTML on delete requests
because the redirection makes it unnecessary. The browser will
send a get to "/" on success. HTML rendering is disabled annotating
the handler function with @ResponseBody and making it void.