Refactor form functions javascript
Extract logic to adjust indexes of items array to its own function.
Extract logic to enable of disable Remove buttons to its own function.
Invoke these functions, and also the one for enabling/disabling submit
buttons, whenever the form is loaded.

This will be useful to reuse the same form and functions to edit a
shipment: the fields in the form will not be empty when loading the
page, therefore we will need to run the logic on document ready