Edit shipment
Created form template, which is filled which info from the shipment
entity to be edited. Template is duplicated from the new shipment
static form.

Shipments are properly edited (not duplicated). For this to work
properly, the form has to send back to the server the shipment id.
This is achieved with a hidden input element in the form template.
Also, getter and setter for the id attribute are needed.

Same rationale behind the items edition: added a hidden field to
send back items ids. But, when a new item is added, remove this
hidden field, as a new item entity has to be created. Also, the
name attribute of the hidden attribute needs to have its array
index adjusted.