Avoid HTML form code duplication
Generate both "New shipment" and "Edit shipment" from the same
mustache template. Create a new handler in the controler to generate
the new shipment form from the template. The handler adds an empty
shipment instance to the model for convenience. Added conditions
in the template to render or not the hidden id fields.

Rename files conveniently